replace ByteString with ByteBuffer in BasicKeyChain
ByteString features good immutability but immutability can be assured
with a bit of care, here, too. The replaced ByteStrings were keys to
maps that never get exposed to other classes and when written into the
map, a copy is being used.
ByteBuffer is probably the most light-weight way of achieving this.